Merge for 0.9.7 release

+++ DisconnectedClient/Program.cs
@@ -1,4 +1,23 @@
-﻿#if !IE
+﻿/*
+ * Copyright 2015-2017 Mohawk College of Applied Arts and Technology
+ * 
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you 
+ * may not use this file except in compliance with the License. You may 
+ * obtain a copy of the License at 
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0 
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the 
+ * License for the specific language governing permissions and limitations under 
+ * the License.
+ * 
+ * User: justi
+ * Date: 2017-4-3
+ */
+#if !IE
 using CefSharp;
 #endif
 using MohawkCollege.Util.Console.Parameters;
@@ -42,10 +61,13 @@ namespace DisconnectedClient
                 Console.WriteLine("Will start in debug mode...");
             if (Program.Parameters.Reset)
             {
-                var appData = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "OpenIZDC");
-                var cData = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "OpenIZDC");
-                if (Directory.Exists(appData)) Directory.Delete(cData, true);
-                if (Directory.Exists(appData)) Directory.Delete(appData, true);
+                if (MessageBox.Show("Are you sure you want to wipe all your data and configuration for the Disconnected Client?", "Confirm Reset", MessageBoxButtons.OKCancel) == DialogResult.OK)
+                {
+                    var appData = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "OpenIZDC");
+                    var cData = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "OpenIZDC");
+                    if (Directory.Exists(appData)) Directory.Delete(cData, true);
+                    if (Directory.Exists(appData)) Directory.Delete(appData, true);
+                }
                 return;
             }
             String[] directory = {
@@ -83,15 +105,13 @@ namespace DisconnectedClient
             // Start up!!!
             try
             {
+
 #if IE
 #else
-                uint x, y;
-                Screen.PrimaryScreen.GetDpi(DpiType.Angular, out x, out y);
-                if (x > 120 || y > 120)
-                    Cef.EnableHighDPISupport();
-                var settings = new CefSettings();
+                var settings = new CefSettings() { UserAgent = "OpenIZEmbedded" };
                 Cef.Initialize(settings, performDependencyCheck: true, browserProcessHandler: null);
 #endif
+
                 Application.EnableVisualStyles();
                 Application.SetCompatibleTextRenderingDefault(false);
 
@@ -119,19 +139,25 @@ namespace DisconnectedClient
                             Application.DoEvents();
                     }
 
-                    main = new frmDisconnectedClient("http://127.0.0.1:9200/org.openiz.core/views/settings/index.html");
+                    if (minims.IsRunning)
+                        main = new frmDisconnectedClient("http://127.0.0.1:9200/org.openiz.core/views/settings/splash.html");
+                    else return;
                 }
-                else
+                else 
                 {
+
 
                     DcApplicationContext.Current.Started += startHandler;
                     while (!started)
                         Application.DoEvents();
+
                     main = new frmDisconnectedClient("http://127.0.0.1:9200/org.openiz.core/splash.html");
                 }
 
                 splash.Close();
-                Application.Run(main);
+
+                if(XamarinApplicationContext.Current.GetService<MiniImsServer>().IsRunning)
+                    Application.Run(main);
